The Python Interface
In the Python script on the right, you can type Python code to solve the exercises. If you hit Run Code or Submit Answer, your python script (script.py) is executed 
and the output is shown in the IPython Shell. Submit Answer checks whether your submission is correct and gives you feedback.

You can hit Run Code and Submit Answer as often as you want. If you're stuck, you can click Get Hint, and ultimately Get Solution.

You can also use the IPython Shell interactively by simply typing commands and hitting Enter. When you work in the shell directly, your code will not be checked for 
correctness so it is a great way to experiment.

Instructions
100 XP

Experiment in the IPython Shell; type 5 / 8, for example.
Add another line of code to the Python script on the top-right (not in the Shell): print(7 + 10).
Hit Submit Answer to execute the Python script and receive feedback.